% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbb_pal.R
\name{fbb_pal}
\alias{fbb_pal}
\title{FBB Color Palettes}
\usage{
fbb_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{named color palette (see the list in fbb_palettes)}

\item{reverse}{(default FALSE) reverses the order of colors}
}
\value{
a ColorRampPalette function
}
\description{
Creates a color palette based on FBB colors
}
\examples{
fbb_pal()

}
