% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Adds the FBB logo to plots}
\usage{
finalise_plot(
  plot_name,
  source_name,
  logo_image_path = file.path(system.file("data", package = "fbb"), "BB_logo_small.png")
)
}
\arguments{
\item{palette}{named color palette (see the list in fbb_palettes)}

\item{reverse}{(default FALSE) reverses the order of colors}
}
\value{
a ColorRampPalette function
}
\description{
Adds the FBB logo to plots
}
\examples{
fbb_pal()

}
