% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fbb.R
\name{theme_fbb}
\alias{theme_fbb}
\title{FBB theme}
\usage{
theme_fbb(
  base_size = 14,
  base_family = "CircularStd",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\value{
a ggplot2 theme
}
\description{
Creates an FBB theme.
    Requires fonts Circular Std regular (# https://github.com/elartix/circular-std/blob/master/fonts/CircularStd-Book.ttfy)
    and bold (https://github.com/elartix/circular-std/blob/master/fonts/CircularStd-Bold.ttf .
    After instlling the fonts, register them in R:
    # install.packages("extrafont")
    # extrafont::font_import()
    # extrafont::loadfonts()
}
\examples{
theme_fbb()

}
